improvement to active setting
Making sure the active setting enables or disables more than just the actionable part of each plugin, so only the commands and absolute basics are available when plugins are disabled.

+++ captcha.ts
@@ -269,43 +269,20 @@ export = C.mount("message", async (ctx: ExtendedContext, next) => {
         }
     }
 
-    const members = ctx.message?.new_chat_members?.filter(
-        (x) => x.username !== ctx.me
-    );
-    if (!members || members.length === 0) {
-        const foundChallenge = activeChallenges.find(
-            (x) => (x.user.id === ctx.from?.id && x.chat === String(ctx.chat?.id))
+    if (settings.active) {
+
+        const members = ctx.message?.new_chat_members?.filter(
+            (x) => x.username !== ctx.me
         );
-        if (!foundChallenge) {
-            return next();
-        }
-        if (Number(ctx.message?.text) === foundChallenge.math[1]) {
-            //logError("Correct Answer: \n" + ctx.message?.text + " " + foundChallenge.math[1] + "\n" + ctx.chat?.id + " " + foundChallenge.chat);
-            clearTimeout(foundChallenge.kickTimeout);
-            activeChallenges.splice(activeChallenges.indexOf(foundChallenge), 1);
-            return Promise.all([
-                ctx.message?.message_id
-                    ? ctx.deleteMessage(ctx.message?.message_id)
-                    : Promise.resolve(true),
-                ctx.deleteMessage(foundChallenge.messageId),
-                foundChallenge.deleteList.forEach(number => {
-                    ctx.deleteMessage(number);
-                }),
-                ctx.replyWithHTML(html`Thank you, ${link(ctx.from)}, please read the rules in our pinned message and enjoy the chat!`),
-                ctx.telegram.restrictChatMember(ctx.chat?.id, ctx.from?.id, verifiedOptions),
-            ]).catch((err) => logError("[captcha] " + err.message));
-        }
-        if (Number(ctx.message?.text) !== foundChallenge.math[1]) {
-            //logError("Wrong Answer: \n" + ctx.message?.text + " " + foundChallenge.math[1] + "\n" + ctx.chat?.id + " " + foundChallenge.chat);
-            if (foundChallenge.attempts >= 1) {
-                let msg;
-                return Promise.all([
-                    foundChallenge.deleteList.push(ctx.message?.message_id),
-                    msg = await ctx.replyWithHTML(html`Sorry, ${link(ctx.from)}, that answer is incorrect, you have ${foundChallenge.attempts} attempts left, try again!`),
-                    foundChallenge.deleteList.push(msg?.message_id),
-                    foundChallenge.attempts = foundChallenge.attempts - 1,
-                ]).catch((err) => logError("[captcha] " + err.message));
-            } else {
+        if (!members || members.length === 0) {
+            const foundChallenge = activeChallenges.find(
+                (x) => (x.user.id === ctx.from?.id && x.chat === String(ctx.chat?.id))
+            );
+            if (!foundChallenge) {
+                return next();
+            }
+            if (Number(ctx.message?.text) === foundChallenge.math[1]) {
+                //logError("Correct Answer: \n" + ctx.message?.text + " " + foundChallenge.math[1] + "\n" + ctx.chat?.id + " " + foundChallenge.chat);
                 clearTimeout(foundChallenge.kickTimeout);
                 activeChallenges.splice(activeChallenges.indexOf(foundChallenge), 1);
                 return Promise.all([
@@ -316,30 +293,59 @@ export = C.mount("message", async (ctx: ExtendedContext, next) => {
                     foundChallenge.deleteList.forEach(number => {
                         ctx.deleteMessage(number);
                     }),
-                    ctx.kickChatMember(ctx.from?.id, Math.floor((Date.now() / 1000) + settings.kickCooldown)),
+                    ctx.replyWithHTML(html`Thank you, ${link(ctx.from)}, please read the rules in our pinned message and enjoy the chat!`),
+                    ctx.telegram.restrictChatMember(ctx.chat?.id, ctx.from?.id, verifiedOptions),
                 ]).catch((err) => logError("[captcha] " + err.message));
             }
+            if (Number(ctx.message?.text) !== foundChallenge.math[1]) {
+                //logError("Wrong Answer: \n" + ctx.message?.text + " " + foundChallenge.math[1] + "\n" + ctx.chat?.id + " " + foundChallenge.chat);
+                if (foundChallenge.attempts >= 1) {
+                    let msg;
+                    return Promise.all([
+                        foundChallenge.deleteList.push(ctx.message?.message_id),
+                        msg = await ctx.replyWithHTML(html`Sorry, ${link(ctx.from)}, that answer is incorrect, you have ${foundChallenge.attempts} attempts left, try again!`),
+                        foundChallenge.deleteList.push(msg?.message_id),
+                        foundChallenge.attempts = foundChallenge.attempts - 1,
+                    ]).catch((err) => logError("[captcha] " + err.message));
+                } else {
+                    clearTimeout(foundChallenge.kickTimeout);
+                    activeChallenges.splice(activeChallenges.indexOf(foundChallenge), 1);
+                    return Promise.all([
+                        ctx.message?.message_id
+                            ? ctx.deleteMessage(ctx.message?.message_id)
+                            : Promise.resolve(true),
+                        ctx.deleteMessage(foundChallenge.messageId),
+                        foundChallenge.deleteList.forEach(number => {
+                            ctx.deleteMessage(number);
+                        }),
+                        ctx.kickChatMember(ctx.from?.id, Math.floor((Date.now() / 1000) + settings.kickCooldown)),
+                    ]).catch((err) => logError("[captcha] " + err.message));
+                }
+            }
+            return settings.strict
+                ? ctx.message?.message_id
+                    ? ctx.deleteMessage(ctx.message?.message_id)
+                    : Promise.resolve(true)
+                : next();
         }
-        return settings.strict
-            ? ctx.message?.message_id
-                ? ctx.deleteMessage(ctx.message?.message_id)
-                : Promise.resolve(true)
-            : next();
-    }
-    return Promise.all(
-        members.map(async (x) => {
-            if (settings.active && !x.is_bot) {
-                const c = challenge(activeChallenges, ctx, x);
-                activeChallenges.push(c);
-                const msg = await ctx.replyWithHTML(
-                    html`Welcome ${link(x)}, please solve the following arithmetic operation in ${(settings.challengeTimeout / 60)} minutes: \n\n
+
+        return Promise.all(
+            members.map(async (x) => {
+                if (!x.is_bot) {
+                    const c = challenge(activeChallenges, ctx, x);
+                    activeChallenges.push(c);
+                    const msg = await ctx.replyWithHTML(
+                        html`Welcome ${link(x)}, please solve the following arithmetic operation in ${(settings.challengeTimeout / 60)} minutes: \n\n
                 <code>${c.math[0]}</code> \n\n
                 Note: Results can be negative, don't forget the (-) sign if so`
-                );
-                c.chat = String(ctx.chat?.id);
-                c.messageId = msg.message_id;
-                ctx.telegram.restrictChatMember(ctx.chat?.id, ctx.from?.id, unverifiedOptions);
-            }
-        })
-    );
+                    );
+                    c.chat = String(ctx.chat?.id);
+                    c.messageId = msg.message_id;
+                    ctx.telegram.restrictChatMember(ctx.chat?.id, ctx.from?.id, unverifiedOptions);
+                }
+            })
+        );
+    } else {
+        return next();
+    }
 });
